Enable PDF export support and refactor image processing
- Add PDF export to buffer method with aspect ratio preservation
- Move related functions to dedicated 'features' and 'generate' folders
- Improve code organization and maintainability for image rendering
