+++ package.json
@@ -19,6 +19,7 @@
     "express": "^4.21.2",
     "express-rate-limit": "^7.5.1",
     "node-cache": "^5.1.2",
+    "pdfkit": "^0.17.1",
     "sharp": "^0.33.5"
   },
   "devDependencies": {
@@ -26,6 +27,7 @@
     "@types/jest": "^29.5.14",
     "@types/mapbox__polyline": "^1.0.5",
     "@types/node": "^22.15.34",
+    "@types/pdfkit": "^0.14.0",
     "@types/supertest": "^6.0.3",
     "cross-env": "^7.0.3",
     "jest": "^29.7.0",
